Move text.dart to widgets test folder
