+++ packages/flutter/test/material/text_field_test.dart
@@ -16,7 +16,7 @@ import 'package:flutter/foundation.dart';
 import 'package:flutter/gestures.dart' show DragStartBehavior, PointerDeviceKind;
 
 import '../rendering/mock_canvas.dart';
-import '../text.dart' show findRenderEditable, globalize, textOffsetToPosition;
+import '../widgets/text.dart' show findRenderEditable, globalize, textOffsetToPosition;
 import '../widgets/semantics_tester.dart';
 import 'feedback_tester.dart';
