+++ packages/flutter/test/material/text_selection_test.dart
@@ -6,7 +6,7 @@ import 'package:flutter_test/flutter_test.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter/rendering.dart';
 import 'package:flutter/widgets.dart';
-import '../text.dart' show findRenderEditable, globalize, textOffsetToPosition;
+import '../widgets/text.dart' show findRenderEditable, globalize, textOffsetToPosition;
 
 void main() {
   group('canSelectAll', () {
